+++ test/t8_cmesh/t8_gtest_compute_first_element.cxx
@@ -135,8 +135,8 @@ TEST_P (t8_gtest_rank_times_global_num_elems_over_size, large_numbers)
    * We use integer division, therefore we store the remainder of each update to
    * prevent rounding errors.
   */
-  uint32_t size = 1;
-  // uint64_t size = 1;
+  // uint32_t size = 1;
+  uint64_t size = 1;
   for (uint32_t isize = 1; isize < size_iter; ++isize) {
     /* The very first result is 1 * 1 / size */
     uint64_t check_result_elem = 1 / size;
@@ -153,7 +153,7 @@ TEST_P (t8_gtest_rank_times_global_num_elems_over_size, large_numbers)
       /* The remainder of the rank update */
       uint64_t rank_remainder = check_result_elem_remain;
       for (uint32_t irank = 1; irank < rank_iter && rank <= size; ++irank) {
-        const uint64_t computed_result = t8_cmesh_get_first_element_of_process (rank, (uint32_t) size, num_elems);
+        const uint64_t computed_result = t8_cmesh_get_first_element_of_process (rank, size, num_elems);
         check_result = (rank == size) ? num_elems : check_result;
 
         ASSERT_EQ (computed_result, check_result)
